Expences card converted to components
