+++ front-end-side/src/Components/ExpensesCard.css
@@ -1,0 +1,2 @@
+body {
+}
