+++ front-end-side/src/index.css
@@ -11,8 +11,8 @@
     /*Statistics bars colors*/
     --food-color: #ff2f39;
     --food-color-shade: #ff27b3;
-    --clothing-color: #8f03ff;
-    --clothing-color-shade: #ba15ff;
+    --clothes-color: #8f03ff;
+    --clothes-color-shade: #ba15ff;
     --entertainment-color: #df7211;
     --entertainment-color-shade: #ffc614;
     --transportation-color: #0067ad;
